move perdiem func to new file
