+++ src/pages/iou/request/step/IOURequestStepSubrate.tsx
@@ -21,7 +21,8 @@ import useThemeStyles from '@hooks/useThemeStyles';
 import {addErrorMessage} from '@libs/ErrorUtils';
 import Navigation from '@libs/Navigation/Navigation';
 import {getPerDiemCustomUnit} from '@libs/PolicyUtils';
-import {addSubrate, getIOURequestPolicyID, removeSubrate, updateSubrate} from '@userActions/IOU';
+import {getIOURequestPolicyID} from '@userActions/IOU';
+import {addSubrate, removeSubrate, updateSubrate} from '@userActions/IOU/PerDiem';
 import CONST from '@src/CONST';
 import type {TranslationPaths} from '@src/languages/types';
 import ONYXKEYS from '@src/ONYXKEYS';
